requirements.txt
requests==2.30.0
python-dotenv==1.0.0
python-telegram-bot==13
vk_api==11.9.9
redis==5.0.0
